Update package dependencies and add 'licenses.txt'
Introduced 'copy-webpack-plugin' in 'package.json' for copying individual files or entire directories to the build directory. This is important for moving files that aren't directly imported by our modules. Also added a 'licenses.txt' in the 'src/txt' directory for declaring the licences of website content and images. Expanded 'webpack.common.js' to include a new instance of 'CopyPlugin' to handle the copying of files from 'src/txt'. Updated yarn.lock as per the updated dependencies.